+++ src/graphql/authentication/authResult.interface..ts
@@ -1,0 +1,14 @@
+import { IUser } from '../../interfaces/IUser.interface';
+
+export interface IAuthResult {
+  user: IUser;
+  access_token: string;
+}
+
+export interface ILoginResult {
+  login: IAuthResult;
+}
+
+export interface ISignupResult {
+  signup: IAuthResult;
+}
