+++ src/pages/SignUpPage/SignUpPage.tsx
@@ -7,9 +7,10 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import { useMutation } from '@apollo/client';
 import { RoutePath } from '../../constants/routeVariables';
 import { schema } from '../../utils/validationSchema';
-import { IFormInput } from '../../interfaces/input/IFormInput.interface';
 import { SIGNUP } from '../../graphql/authentication/mutation';
 import { authService } from '../../graphql/authentication/authService';
+import { ISignupResult } from '../../graphql/authentication/authResult.interface.';
+import { IFormInput } from './formInput.interface';
 import {
   ButtonSubmitForm,
   FormSign,
@@ -21,7 +22,7 @@ import {
 
 const SignUpPage = () => {
   const [hiddenPassword, setHiddenPassword] = useState(true);
-  const [signUp] = useMutation(SIGNUP);
+  const [signUp] = useMutation<ISignupResult>(SIGNUP);
   const navigate = useNavigate();
 
   const {
